requirements.txt
streamlit==1.29.0
pycaret==3.1.0
openai==1.3.6
langchain==0.0.350
pinecone-client==2.2.4
tiktoken==0.5.2
protobuf==3.20.0
fpdf==1.7.2